requirements.txt
opencv-python
numpy
tqdm
onnxruntime-gpu==1.18.0

